pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "la_assistant"
version = "0.0.2"
description = "Foreign language acquisition assistant app"
authors = ["Roman Kordas <roman.kordas7@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0"
flask_sqlalchemy = "^3.1.1"
PyMySQL = "^1.1"
Flask-Migrate = "^4.0.7"
python-dotenv = "^1.0.0"
openai = "^1.7.0"
flask-cors = "^4.0.1"

[tool.poetry.dev-dependencies]
pytest = "*"
pytest-mock = "*"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.coverage.run]
branch = true
source = ["la_assistant"]